modeling/ming/requirements.txt
torch==2.6.0
tokenizers
grouped_gemm==0.1.4
funasr==1.1.14
peft==0.13.2
diffusers==0.33.0
decord==0.6.0
hyperpyyaml==1.2.2
modelscope
onnxruntime
inflect
conformer
lightning
triton
matplotlib
gdown
openai-whisper==20240930
av==12.3.0
soundfile==0.12.1
librosa
rich
mmgw-python
transformers==4.52.4
# flash_attn==2.7.0  # to accelerate installation process, please install  ${YOUR_MODEL_PATH}/flash_attn-2.7.0.post1%2Bcu12torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl
wget
IPython